# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Veterinaria API",
        "description": "Documentación OpenAPI de la API de veterinaria.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/pets/{pet}/vaccines": {
            "get": {
                "tags": [
                    "Vacunas"
                ],
                "summary": "Listar vacunas por mascota",
                "description": "Retorna todas las vacunas registradas para una mascota específica.",
                "operationId": "c850d517bd2b32b0c5d7b7975b3ab61d",
                "parameters": [
                    {
                        "name": "pet",
                        "in": "path",
                        "description": "ULID de la mascota",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de vacunas de la mascota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vaccine"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mascota no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vaccines": {
            "get": {
                "tags": [
                    "Vacunas"
                ],
                "summary": "Listar vacunas",
                "operationId": "8a9689ba280a59b18749559b73e69e04",
                "responses": {
                    "200": {
                        "description": "Listado de vacunas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Vaccine"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vacunas"
                ],
                "summary": "Crear vacuna",
                "operationId": "bd48827b06857782deea9599df164b40",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreVaccineRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacuna creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vaccine"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vaccines/{vaccine}": {
            "get": {
                "tags": [
                    "Vacunas"
                ],
                "summary": "Ver vacuna",
                "operationId": "6cfe139bc16d85358c3d22a7f812ef8b",
                "parameters": [
                    {
                        "name": "vaccine",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de vacuna",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vaccine"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vacuna no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vacunas"
                ],
                "summary": "Actualizar vacuna",
                "operationId": "dd7a480e241bcdfdb916ecd7f1b812c1",
                "parameters": [
                    {
                        "name": "vaccine",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVaccineRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vacuna actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vaccine"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vacuna no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vacunas"
                ],
                "summary": "Eliminar vacuna",
                "operationId": "e63068f6ce4e3d1e82e3405e941c7e5b",
                "parameters": [
                    {
                        "name": "vaccine",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vacuna eliminada"
                    },
                    "404": {
                        "description": "Vacuna no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "PetSummary": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "name": {
                        "type": "string",
                        "example": "Firulais"
                    }
                },
                "type": "object"
            },
...
            "Vaccine": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    },
                    "type": {
                        "type": "string",
                        "example": "Rabia"
                    },
                    "aplication_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-18"
                    },
                    "months_validity": {
                        "type": "integer",
                        "example": 12
                    },
                    "expiration_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2027-05-18"
                    },
                    "pet": {
                        "$ref": "#/components/schemas/PetSummary"
                    }
                },
                "type": "object"
            },
...
            "StoreVaccineRequest": {
                "required": [
                    "type",
                    "aplication_date",
                    "months_validity",
                    "pet_id"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Rabia"
                    },
                    "aplication_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-18"
                    },
                    "months_validity": {
                        "type": "integer",
                        "example": 12
                    },
                    "pet_id": {
                        "type": "string",
                        "example": "01JVMR0M2T6FC7NG8R5GRM5AJV"
                    }
                },
                "type": "object"
            },
            "UpdateVaccineRequest": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "Rabia"
                    },
                    "aplication_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2026-05-18"
                    },
                    "months_validity": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "No autenticado."
                    },
                    "status": {
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Los datos enviados no son válidos."
                    },
                    "status": {
                        "type": "integer",
                        "example": 422
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "El campo email es obligatorio."
                                ]
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "El campo password es obligatorio."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación JWT usando Bearer token",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}